Check SIGSEGV better
