+++ bin/utest/mmap.c
@@ -94,15 +94,26 @@ static void munmap_good(void) {
   assert(result == 0);
 }
 
-/* Don't call this function in this module */
 int test_munmap_sigsegv(void) {
+  setup_sigsegv_sigaction();
+
   void *addr = mmap_anon_prw(NULL, 0x4000);
+  volatile int *ptr = (volatile int *)(addr + 0x2000);
+
   munmap(addr, 0x4000);
 
-  /* Try to access freed memory. It should raise SIGSEGV */
-  int data = *((volatile int *)(addr + 0x2000));
-  (void)data;
-  return 1;
+  if (sigsetjmp(return_to, 1) == 0) {
+    /* Try to access freed memory. It should raise SIGSEGV */
+    (void)*ptr;
+    assert(0);
+  }
+
+  /* Check if SIGSEGV was handled correctly */
+  assert(sigsegv_address == ptr);
+  assert(sigsegv_code = SEGV_MAPERR);
+
+  signal(SIGSEGV, SIG_DFL);
+  return 0;
 }
 
 int test_mmap(void) {
@@ -146,33 +157,26 @@ int test_munmap(void) {
   return 0;
 }
 
-static volatile int sigsegv_handled = 0;
-static jmp_buf return_to;
-static void sigsegv_handler(int signo) {
-  sigsegv_handled++;
-  siglongjmp(return_to, 1);
-}
-
 #define NPAGES 8
 
 int test_mmap_prot_none(void) {
-  signal(SIGSEGV, sigsegv_handler);
+  setup_sigsegv_sigaction();
 
   size_t pgsz = getpagesize();
   size_t size = pgsz * NPAGES;
   volatile void *addr = mmap_anon_priv(NULL, size, PROT_NONE);
   assert(addr != MAP_FAILED);
 
-  sigsegv_handled = 0;
-
   for (int i = 0; i < NPAGES; i++) {
     volatile uint8_t *ptr = addr + i * pgsz;
     if (sigsetjmp(return_to, 1) == 0) {
       assert(*ptr == 0);
     }
-  }
 
-  assert(sigsegv_handled == NPAGES);
+    /* Check if SIGSEGV was handled correctly */
+    assert(sigsegv_address == ptr);
+    assert(sigsegv_code = SEGV_ACCERR);
+  }
 
   /* restore original behavior */
   signal(SIGSEGV, SIG_DFL);
@@ -181,7 +185,7 @@ int test_mmap_prot_none(void) {
 }
 
 int test_mmap_prot_read(void) {
-  signal(SIGSEGV, sigsegv_handler);
+  setup_sigsegv_sigaction();
 
   size_t pgsz = getpagesize();
   size_t size = pgsz * NPAGES;
@@ -192,17 +196,18 @@ int test_mmap_prot_read(void) {
   for (size_t i = 0; i < size / sizeof(uint32_t); i++)
     assert(((uint32_t *)addr)[i] == 0);
 
-  sigsegv_handled = 0;
-
   for (int i = 0; i < NPAGES; i++) {
     volatile uint8_t *ptr = addr + i * pgsz;
     if (sigsetjmp(return_to, 1) == 0) {
       *ptr = 42;
     }
+    /* Check if nothing changed */
     assert(*ptr == 0);
-  }
 
-  assert(sigsegv_handled == NPAGES);
+    /* Check if SIGSEGV was handled correctly */
+    assert(sigsegv_address == ptr);
+    assert(sigsegv_code = SEGV_ACCERR);
+  }
 
   /* restore original behavior */
   signal(SIGSEGV, SIG_DFL);
